Silence -Wunused-parameter honestly; last -Wextra nits; drop all -Wno-* flags
- 46 unused parameters (virtual callback stubs, placement operator new
  size, empty inline hooks): parameter names commented out at the
  declaration, the standard idiom that keeps the signature readable.
- time.cpp: TimeData aggregate init spelled out in full instead of
  { 3 } (-Wmissing-field-initializers); timerfd read() in NanoSleep now
  breaks out on error instead of spinning.
- snprintf.c / strptime.c: signed/unsigned comparison fixes in the
  vendored C (cast the constant side, no behaviour change).

gcc_wflags is now plain -Wall -Wextra.  The repo builds with zero
warnings from its own sources; remaining LTO-time notes originate in
sibling raikv/raimd headers inlined across the link.

+++ src/util/strptime.c
@@ -273,7 +273,7 @@ rai_strptime(const char *buf, const char *fmt, struct tm *tm)
 
                 case 'A':
                 case 'a':
-                        for (i = 0; i < asizeof(rai_En_US.weekday_names); i++) {
+                        for (i = 0; i < (int) asizeof(rai_En_US.weekday_names); i++) {
                                 len = strlen(rai_En_US.weekday_names[i]);
                                 if (strncasecmp(buf,
                                                 rai_En_US.weekday_names[i],
@@ -315,7 +315,7 @@ rai_strptime(const char *buf, const char *fmt, struct tm *tm)
                 case 'B':
                 case 'b':
                 case 'h':
-                        for (i = 0; i < asizeof(rai_En_US.month_names); i++) {
+                        for (i = 0; i < (int) asizeof(rai_En_US.month_names); i++) {
                                 len = strlen(rai_En_US.month_names[i]);
                                 if (strncasecmp(buf,
                                                 rai_En_US.month_names[i],
